% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test.R
\docType{data}
\name{data_test}
\alias{data_test}
\title{test dataset}
\format{a data frame
\describe{
\item{year}{year of transaction}
\item{model}{model sold}
\item{volume}{volume sold, unit}
}}
\source{
source data for the agregation process of the products
}
\usage{
data_test
}
\description{
generated in raw-data.
}
\keyword{datasets}
